Split native action precision equality
